fix up all the tests
